propagacion de cambios a los test repositorios

+++ src/test/java/ar/edu/utn/frba/dds/models/repositories/contribucion/CuidadoHeladerasRepositoryTest.java
@@ -36,7 +36,7 @@ class CuidadoHeladerasRepositoryTest {
 
   @Test
   void testInsertContribucion() throws RepositoryException {
-    int id = repositorio.insert(contribucion);
+    Long id = repositorio.insert(contribucion);
 
     assertEquals(1, id);
     assertEquals(1, contribucion.getId());
