+++ src/test/java/ar/edu/utn/frba/dds/models/repositories/contribucion/DonacionViandasRepositoryTest.java
@@ -91,10 +91,10 @@ class DonacionViandasRepositoryTest {
 
   @Test
   void testInsertarDonacion() throws RepositoryException {
-    int id = repositorio.insert(donacion);
+    Long id = repositorio.insert(donacion);
 
-    assertEquals(1, id);
-    assertEquals(1, donacion.getId());
+    assertEquals(1L, id);
+    assertEquals(1L, donacion.getId());
   }
 
   @Test
